// FRONTEND/ZKDEX/src/uikit/components/Footer/types.ts
import { Language } from '../LangSelector/types'
import { FlexProps } from '../Box'

export type FooterLinkType = {
  label: string
  items: { label: string; href?: string; isHighlighted?: boolean }[]
}

export type FooterProps = {
  items: FooterLinkType[]
  buyCakeLabel: string
  isDark: boolean
  toggleTheme: (isDark: boolean) => void
  cakePriceUsd?: number
  currentLang: string
  langs: Language[]
  setLang: (lang: Language) => void
} & FlexProps

export type SocialLink = {
  label: string
  icon: string
  href: string
  items?: SocialLink[]
}